Core/Player: fix phase update reward quest

+++ src/server/shared/Cryptography/ARC4.cpp
@@ -26,7 +26,7 @@ ARC4::ARC4(uint32 len) : m_ctx(EVP_CIPHER_CTX_new())
     EVP_CIPHER_CTX_set_key_length(m_ctx, len);
 }
 
-ARC4::ARC4(uint8 *seed, uint8 len) : m_ctx()
+ARC4::ARC4(uint8 *seed, uint32 len) : m_ctx()
 {
     EVP_CIPHER_CTX_init(m_ctx);
     EVP_EncryptInit_ex(m_ctx, EVP_rc4(), NULL, NULL, NULL);
